+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Master
+
+- Improved `ESUniqueNumericIdentifier()` function.
+
 ## 3.18.4 (2019-08-30)
 
 - Improved `ESUniqueNumericIdentifier()` function.